+++ src/selectize.js
@@ -1484,7 +1484,7 @@ $.extend(Selectize.prototype, {
 				if (!self.isPending) {
 					$option = self.getOption(value);
 					value_next = self.getAdjacentOption($option, 1).attr('data-value');
-					self.refreshOptions(self.isFocused && inputMode !== 'single');
+					self.refreshOptions(self.isFocused && inputMode !== 'single' && self.settings.openAfterItemRemove);
 					if (value_next) {
 						self.setActiveOption(self.getOption(value_next));
 					}
@@ -1876,7 +1876,7 @@ $.extend(Selectize.prototype, {
 
 		self.showInput();
 		self.positionDropdown();
-		self.refreshOptions(true);
+		self.refreshOptions(self.settings.openAfterItemRemove);
 
 		// select previous option
 		if (option_select) {